add AuctionService
